Add DB RPC API support to client
We add support for the StoreService
to use the newly added plugin.Driver interface to make
SQL queries.

This allows all plugins using the plugin-api
to route all SQL queries via Mattermost server
using the central connection pool, and avoid
having to open new DB connections.

This is implemented on the server via a database
driver which acts as an RPC wrapper, giving clients
a zero-cost abstraction on which to run their SQL queries.

All clients get a *sql.DB object as normal, but under
the hood they are routed via RPC.

https://focalboard-community.octo.mattermost.com/workspace/zyoahc9uapdn3xdptac6jb69ic?id=285b80a3-257d-41f6-8cf4-ed80ca9d92e5&v=495cdb4d-c13a-4992-8eb9-80cfee2819a4&c=c7386db7-65fd-469b-8bcf-8dc8f8e61e4f